Implement fetch  and display screens
First created a content interactor tests to both understand the rest
api and implement the interaction with the rest api.

Wrote a simple instrumentation test to cover the fetch screen
functionality, and implemented FetchContentActivity along with its
content presenter.

Implemented DisplayContentActivity along with its adapter
and view holders to show the fetched data(seconds). Finally use a
timer observable to cover the countdown mechanism and update the
views displaying the data.